+++ antares/shared/src/main/ch/scorpion/antares/model/memory/CompressedMemoryDump.kt
@@ -27,7 +27,7 @@ object CompressedMemoryDump {
 	 */
 	fun write(memory: Memory, bitWidth: BitWidth): String {
 		val builder = StringBuilder()
-		val mask: Long = BitOperation.power(bitWidth.width.toLong()) - 1L
+		val mask: Long = BitOperation.power(bitWidth.width.toByte()) - 1L
 		val length = Math.max(2, bitWidth.width / 4)
 
 		val cellIter = ZeroFiller(memory.getNonZeroCells())
